RIO-7492: MP4 file parser and parser node modified to support a very large value of timescale

+++ baselibs/media_data_structures/src/pv_gau.h
@@ -39,7 +39,7 @@ struct SimpleBufferFragGroup
 struct MediaMetaInfo
 {
     uint32  len;  //size of the frame
-    uint32  ts;   //time stamp
+    uint64  ts;   //time stamp
     uint32  ts_delta; // timestamp delta from the previous timestamp
     uint32  sample_info;  //4 uint8 information together
     bool  dropFlag;
